211224_07

+++ app/Http/Controllers/shop/BaesongjiController.php
@@ -179,10 +179,28 @@ exit;
         $id         = $request->input('num');
         $user_id    = Auth::user()->user_id;
 
-        DB::table('baesongjis')->where([['id', $id],['user_id',$user_id]])->delete();   //row 삭제
+        //DB::table('baesongjis')->where([['id', $id],['user_id',$user_id]])->delete();   //row 삭제
+        $baesongji_chk = DB::table('baesongjis')->where([['id', $id],['user_id',$user_id]])->first();
 
-        echo "ok";
-        exit;
+        if($baesongji_chk->ad_default == 1){
+            echo "default_no_del";
+            exit;
+        }else{
+            $up_result = DB::table('users')->where('user_id', Auth::user()->user_id)->update([
+                'user_zip'          => '',
+                'user_addr1'        => '',
+                'user_addr2'        => '',
+                'user_addr3'        => '',
+                'user_addr_jibeon'  => '',
+            ]);
+
+            DB::table('baesongjis')->where([['id', $id],['user_id',$user_id]])->delete();   //row 삭제
+
+            echo "ok";
+            exit;
+        }
+
+
     }
 
     public function ajax_ordersendcost(Request $request)
